Ajustes en las respuestas de usuarios, roles y permisos

+++ app/Http/Controllers/Api/Admin/RoleController.php
@@ -7,8 +7,9 @@ use Illuminate\Http\Request;
 use App\Http\Requests\StoreRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
 use App\Http\Requests\SyncPermissionsRequest;
-use Spatie\Permission\Models\Role;
-use Spatie\Permission\Models\Permission;
+use App\Http\Resources\RolesResource;
+use App\Http\Resources\ShowRoleResource;
+use App\Models\Role;
 use App\Traits\ApiResponse;
 
 class RoleController extends Controller
@@ -17,14 +18,28 @@ class RoleController extends Controller
     
     use ApiResponse;
 
-    public function index()
+    public function index(Request $request)
     {
-        $roles = Role::with('permissions')->get();
+        $search = $request->query('search');
 
+        $roles = Role::when($search, function ($query, $search) {
+            $query->where('name', 'like', "%{$search}%");
+        })->orderBy('name')->paginate(10);
+
+        $data = RolesResource::collection($roles->items());
+    
         return $this->successResponse(
             'Roles obtenidos correctamente.',
-            $roles,
-            200
+            $data,
+            200,
+            [
+                'pagination' => [
+                    'total'       => $roles->total(),
+                    'perPage'     => $roles->perPage(),
+                    'currentPage' => $roles->currentPage(),
+                    'lastPage'    => $roles->lastPage(),
+                ]
+            ]
         );
     }
 
@@ -36,7 +51,7 @@ class RoleController extends Controller
 
     public function show(Role $role)
     {
-        return response()->json($role->load('permissions'));
+        return $this->successResponse('Rol encontrado.', new  ShowRoleResource($role));
     }
 
     public function update(UpdateRoleRequest $request, Role $role)
@@ -45,9 +60,31 @@ class RoleController extends Controller
         return response()->json($role);
     }
 
+    public function activate(Role $role)
+    {
+        $role->update(['active' => !$role->active]);
+
+        $data = [
+            'id' => $role->id,
+            'active' => $role->active,
+        ];
+
+        $message = $role->active ? 'Rol activado correctamente.' : 'Rol desactivado correctamente.';
+
+        return $this->successResponse($message, $data);
+    }
+
     public function destroy(Role $role)
     {
+        
+        if (!$role->canBeDeleted()) {
+
+            return $this->errorResponse('Hubo un error al eliminar el rol.', ['name' => ['Este rol está en uso. Desactívalo en lugar de borrarlo.']], 422);
+
+        }
+        
         $role->delete();
+        
         return response()->json(['message' => 'Role eliminado correctamente']);
     }
